function to en/disable argument count check
Sometimes I need to pass more arguments in the RPC call, check the basic parameters in PHP and then work with the rest of JSON data in the plpgsql. Then I need to disable $nbParams > $nbMaxParams check.   When I pass more arguments, the check nbParams < $nbRequiredParams  is useless, it still doesn't guarantee that you are passing all required params (this is done in getNamedArguments()), so you can also disable this check ...